fix: deep parse at-rule preludes
